fix(feedback): allow signed bridge requests
